feat(netid): Add package for generating unique network IDs
Create a new package netid in the pkg/ directory that generates an 8-character unique network ID for each machine based on the hash of its MAC address. This package can be used to uniquely identify machines on a network.

Add a small test in netid_test.go to ensure that the package is working as expected.