refactor(ecdsa): rename package and functions according to Go conventions
Renamed the ecdsakeypair package to ecdsa/ and the keypair.go file inside the package to key.go. Also renamed the GetPublicKey and GetPrivateKey functions to PublicKey and PrivateKey respectively. The package of the test was also renamed accordingly. These changes were made to follow the Go conventions as suggested in the Effective Go documentation.