refactor: remove .DS_Store file and remove VerifySignature from ecdsaKeyPair
This commit includes two changes. First, a .DS_Store file that was present in internal/example has been removed. This file was accidentally committed by Jordan. Second, the VerifySignature function has been removed from the ecdsaKeyPair struct and the KeyPair interface. This change was made because the ecdsaKeyPair struct should not be needed for verifying a signature.