feat: add discovery package and bootstrap node for network analysis
This commit adds the internal discovery package with Kademlia DHT and mDNS. It also develops a simple bootstrap node that listens on port 1211 and subscribes to a pubsub topic SecuraChainNetworkAnalytics. The node only listens to messages sent on this topic and displays the messages on a web page on port 1212. This is intended to assist in the development of the project by providing network analysis and debugging capabilities.