refactor: simplify bootstrap node and update project configuration
This commit makes several changes. First, it removes the "SecuraChainNetworkAnalytics" pubsub and the web server from the bootstrap node to keep it simple. Second, it updates the golangci-lint configuration file to allow the GitHub of new dependencies. Finally, it adds the bin directory, which contains the project's executables, to the .gitignore file.