feat(core): Add core package to internal and block.go
This commit introduces a new package, , within the  directory. The  package is designed to contain the fundamental elements of the blockchain, such as blocks, transactions, block validation, etc.

Additionally, a new file, , has been added to the  package. This file defines an initial version of a block structure and includes several related functions.