feat(client): Initiate Fyne tests for user client GUI
This commit marks the beginning of our tests with the Fyne library for the graphical user interface (GUI) of the user client node. The changes are made on the dedicated client branch.

These initial tests lay the groundwork for the development of the user client GUI using Fyne.

BREAKING CHANGE: Initiate Fyne tests for user client GUI