feat(utils): Add ComputeFileChecksum function in checksum.go
This commit introduces the ComputeFileChecksum function in a new file named checksum.go within the pkg/utils package.

The changes include:

- Creation of checksum.go in the pkg/utils package.
- Addition of the ComputeFileChecksum function which computes the checksum of a given file.

This new function enhances the utility of the package by providing a way to compute file checksums, which can be useful in various scenarios such as file integrity checks.